# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "老年人用药管理系统",
    "description": "老人用药管理智能助手后端 API",
    "version": "2.33.0"
  },
  "paths": {
...
    "/api/v1/public/device/register": {
      "post": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Register Device",
        "description": "设备注册/心跳上报\n\n查找逻辑：\n1. 优先按 User.device_id 查找（家属已绑定到真实老人）\n2. 回退按 User.username == device_id 查找（旧虚拟用户）\n3. 都找不到则创建虚拟用户（兼容老人端开机即用场景，待家属后续绑定）\n\n:return: {\"status\": \"ok\", \"user_id\": int}；首次注册额外返回 \"device_token\"",
        "operationId": "register_device_api_v1_public_device_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/offline": {
      "post": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Device Offline",
        "description": "设备主动下线通知\n\n设备正常退出（SIGINT/SIGTERM/进程关闭）时调用，将 last_heartbeat_at 置为很早的时间，\n使 is_online 立即为 false，避免子女端在心跳超时窗口内看到虚假的\"在线\"状态。\n注意：掉电/SIGKILL 等异常退出仍需依赖心跳超时判定。需校验 X-Device-Token。",
        "operationId": "device_offline_api_v1_public_device_offline_post",
        "parameters": [
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceOffline"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/message": {
      "post": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Device Message",
        "description": "接收设备上报消息（校验 device_id 与 X-Device-Token）",
        "operationId": "device_message_api_v1_public_device_message_post",
        "parameters": [
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/upload": {
      "post": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Device Upload",
        "description": "接收设备上传的服药照片（base64 解码后落盘，HuskyLens 采集）",
        "operationId": "device_upload_api_v1_public_device_upload_post",
        "parameters": [
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/status/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Status",
        "description": "获取设备状态信息（供子女端查询，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_status_api_v1_public_device_status__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/ai/ask": {
      "post": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Ai Ask",
        "description": "AI问答（设备端，基于 IP 限流，每分钟每 IP 最多 10 次）\n\n若提供 device_id，需校验 X-Device-Token。",
        "operationId": "ai_ask_api_v1_public_ai_ask_post",
        "parameters": [
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__v1__endpoints__public__AIQuestion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/check/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Check Device",
        "description": "检查设备是否已注册（供子女端绑定时校验，仅返回 exists，不泄露敏感信息）\n\n与设备公开接口一致读取 X-Device-Token 头；仅做存在性查询、不强制校验令牌——\n绑定流程发生在绑定前，调用方此时可能尚未持有可校验的令牌，且本端点只返回\nexists 布尔、不泄露敏感信息。登录态由 family_monitor 的 Web 路由层 require_login 保证。\n\n查找逻辑：优先 device_id 字段，回退 username。",
        "operationId": "check_device_api_v1_public_device_check__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/schedule/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Schedule",
        "description": "获取设备的用药计划（供老人端每分钟轮询，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_schedule_api_v1_public_device_schedule__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/medication_plan": {
      "post": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Set Device Medication Plan",
        "description": "家属通过设备ID设置用药计划（校验 device_id 与 X-Device-Token）",
        "operationId": "set_device_medication_plan_api_v1_public_device_medication_plan_post",
        "parameters": [
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyMedicationPlan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/plans/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Plans",
        "description": "获取设备的所有用药计划（供子女端查看，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_plans_api_v1_public_device_plans__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/records/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Records",
        "description": "获取设备的服药记录（供子女端 BFF 调用，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_records_api_v1_public_device_records__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/device/chat_history/{device_id}": {
      "get": {
        "tags": [
          "设备公开接口"
        ],
        "summary": "Get Device Chat History",
        "description": "获取设备相关的聊天历史（供子女端 BFF 调用，校验 device_id 与 X-Device-Token）",
        "operationId": "get_device_chat_history_api_v1_public_device_chat_history__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-Device-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceMessage": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type",
            "default": "info"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "DeviceMessage",
        "description": "设备消息"
      },
      "DeviceOffline": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "DeviceOffline",
        "description": "设备主动下线通知"
      },
      "DeviceRegister": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          }
        },
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "DeviceRegister",
        "description": "设备注册"
      },
      "DeviceUpload": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "image_base64": {
            "type": "string",
            "title": "Image Base64"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "image_base64"
        ],
        "title": "DeviceUpload",
        "description": "设备上传图片（base64 编码）"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__api__v1__endpoints__public__AIQuestion": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AIQuestion",
        "description": "AI提问"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}